feat: remove unused files and move TreeMap & AreaChart into components folder
