+++ src/components/AreaChart.tsx
