--- src/not-used/Empty.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Empty: React.FC = () => {
-	return <section id="" />;
-};
-
-export default Empty;
