--- src/not-used/inno-ausgaben.json
@@ -1,77 +1,0 @@
-{
-	"nahrung": {
-		"2023": 29,
-		"2022": 48,
-		"delta": -19
-	},
-	"pharma": {
-		"2023": 1588,
-		"2022": 1524,
-		"delta": 64
-	},
-	"holz": {
-		"2023": 57,
-		"2022": 113,
-		"delta": -56
-	},
-	"metall": {
-		"2023": 71,
-		"2022": 88,
-		"delta": -17
-	},
-	"elektroindustrie": {
-		"2023": 966,
-		"2022": 864,
-		"delta": 102
-	},
-	"fahrzeugbau": {
-		"2023": 949,
-		"2022": 595,
-		"delta": 354
-	},
-	"sonstige_konsumgueter": {
-		"2023": 47,
-		"2022": 77,
-		"delta": -30
-	},
-	"energie": {
-		"2023": 201,
-		"2022": 220,
-		"delta": -19
-	},
-	"telekommunikation": {
-		"2023": 273,
-		"2022": 316,
-		"delta": -43
-	},
-	"software": {
-		"2023": 1088,
-		"2022": 977,
-		"delta": 111
-	},
-	"finanz": {
-		"2023": 266,
-		"2022": 245,
-		"delta": 21
-	},
-	"unternehmensberatung": {
-		"2023": 102,
-		"2022": 124,
-		"delta": -22
-	},
-	"architektur": {
-		"2023": 122,
-		"2022": 93,
-		"delta": 29
-	},
-	"forschung": {
-		"2023": 239,
-		"2022": 269,
-		"delta": -30
-	},
-	"kreativ": {
-		"2023": 188,
-		"2022": 56,
-		"delta": 132
-	}
-}
